* move commanline handling to own file
